Add example definition of functions handleOnRedirect and transactionReference
It is a little bit confusing for people who are getting started with flutterwave to know what to put in `transactionReference` and `handleOnRedirect` functions. So I added an example on how those functions can be implemented. The reason for this is that for me It took about 3 days to discover that I have to define those functions for my I own. I checked the whole SDK and even contactated the support team but couldn't find those functions definitions.